updated script example in README to fix variable being escaped
When the DEVICE_MODEL_CURL variable is declared, the second DEVICE_IDENTIFIER after grep is being sent as a literal string because
the '$' is being escaped with backslash. This causes DEVICE_MODEL and DEVICE_YEAR to be set to unknown and N/A if the mac model in
model.txt has two identifiers, and it searches for the second identifier. For example, MacBook Pro (Retina, 15-inch, Mid 2014)|MacBookPro11,2 MacBookPro11,3|2014.
If the identifier is MacBookPro11,3 then unknown is returned because grep is searching the literal string '$DEVICE_IDENTIFIER' and not the value of the variable.